# apps/runtime/src/handlers/logs.rs
//! Log streaming endpoint using Server-Sent Events (SSE).

use axum::{
    Json,
    extract::{Path, Query, State},
    http::HeaderMap,
    response::{
        IntoResponse,
        sse::{Event, KeepAlive, Sse},
    },
};
use serde::Deserialize;
use std::convert::Infallible;
use tokio_stream::StreamExt as _;
use utoipa::{IntoParams, OpenApi, ToSchema};

use crate::services::custom_bots::CUSTOM_BOT_DEPLOYMENT_PREFIX;
use crate::{
    handlers::{
        auth::{ensure_guild_admin, require_identity},
        custom_bots::{ensure_feature_enabled, ensure_scope_owner},
        error::ApiError,
        response::{ApiEventStream, ApiJson},
    },
    log_sink::{self, LogEntry},
    state::AppState,
};

async fn ensure_log_scope_access(
    state: &AppState,
    identity: &crate::handlers::auth::IdentityContext,
    scope_id: &str,
) -> Result<(), ApiError> {
    if scope_id.starts_with(CUSTOM_BOT_DEPLOYMENT_PREFIX) {
        ensure_feature_enabled(state, identity).await?;
        ensure_scope_owner(state, identity, scope_id).await?;
        Ok(())
    } else {
        ensure_guild_admin(state, identity, scope_id).await
    }
}

#[derive(OpenApi)]
#[openapi(
    paths(get_guild_logs, stream_guild_logs),
    components(schemas(LogEntry, LogsQuery))
)]
pub struct LogsApi;

#[derive(Debug, Deserialize, IntoParams, ToSchema)]
pub struct LogsQuery {
    /// Maximum number of log entries to return (default 100, max 1000).
    #[serde(default = "default_limit")]
    pub limit: usize,
}

fn default_limit() -> usize {
    100
}

/// Get recent logs for a specific guild.
#[utoipa::path(
    get,
    path = "/{guild_id}",
    params(
        ("guild_id" = String, Path, description = "Guild ID"),
        LogsQuery
    ),
    summary = "List guild logs",
    description = "Returns recent log entries for a specific guild.",
    responses(
        (status = 200, description = "Recent log entries for guild", body = Vec<LogEntry>)
    ),
    tag = "Logs"
)]
pub async fn get_guild_logs(
    State(state): State<AppState>,
    headers: HeaderMap,
    Path(guild_id): Path<String>,
    Query(query): Query<LogsQuery>,
) -> Result<ApiJson<Vec<LogEntry>>, ApiError> {
    let identity = require_identity(&state, &headers).await?;
    ensure_log_scope_access(&state, &identity, &guild_id).await?;
    let limit = query.limit.min(1000);
    let logs = log_sink::log_sink().recent_for_guild(&guild_id, limit);
    Ok(ApiJson(Json(logs)))
}

/// Stream logs for a specific guild via Server-Sent Events.
#[utoipa::path(
    get,
    path = "/{guild_id}/stream",
    params(("guild_id" = String, Path, description = "Guild ID or User Bot runtime scope ID")),
    summary = "Stream runtime logs",
    description = "Streams new runtime log entries as Server-Sent Events. The caller must manage the guild or own the addressed User Bot runtime scope. Existing entries are available from the non-streaming logs endpoint.",
    responses(
        (status = 200, description = "Server-Sent Event stream containing runtime log entries", content_type = "text/event-stream")
    ),
    tag = "Logs"
)]
pub async fn stream_guild_logs(
    State(state): State<AppState>,
    headers: HeaderMap,
    Path(guild_id): Path<String>,
) -> Result<ApiEventStream, ApiError> {
    let identity = require_identity(&state, &headers).await?;
    ensure_log_scope_access(&state, &identity, &guild_id).await?;
    let receiver = log_sink::log_sink().subscribe();

    let stream = tokio_stream::wrappers::BroadcastStream::new(receiver).filter_map(move |result| {
        match result {
            Ok(entry) => {
                // Filter to only this guild
                if entry.guild_id.as_deref() == Some(guild_id.as_str()) {
                    let json = serde_json::to_string(&entry).ok()?;
                    Some(Ok::<_, Infallible>(Event::default().data(json)))
                } else {
                    None
                }
            }
            Err(_) => None,
        }
    });

    Ok(ApiEventStream(
        Sse::new(stream)
            .keep_alive(KeepAlive::default())
            .into_response(),
    ))
}