# openapi/flora.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flora API",
    "description": "Manages @Flora guild deployments, User Bots, Server Custom Bots, authentication, builds, secrets, key-value stores, logs, and operator orchestration.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "flora API base path"
    }
  ],
  "paths": {
...
    "/guilds/": {
      "get": {
        "tags": ["Guilds"],
        "summary": "List manageable guilds",
        "description": "Returns Discord guilds where the authenticated user has Administrator or Manage Server permission and @Flora is currently installed. Requires an interactive Discord session.",
        "operationId": "list_guilds_handler",
        "responses": {
          "200": {
            "description": "Guilds available for deployment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GuildResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/logs/{guild_id}": {
      "get": {
        "tags": ["Logs"],
        "summary": "List guild logs",
        "description": "Returns recent log entries for a specific guild.",
        "operationId": "get_guild_logs",
        "parameters": [
          {
            "name": "guild_id",
            "in": "path",
            "description": "Guild ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of log entries to return (default 100, max 1000).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent log entries for guild",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEntry"
                  }
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/logs/{guild_id}/stream": {
      "get": {
        "tags": ["Logs"],
        "summary": "Stream runtime logs",
        "description": "Streams new runtime log entries as Server-Sent Events. The caller must manage the guild or own the addressed User Bot runtime scope. Existing entries are available from the non-streaming logs endpoint.",
        "operationId": "stream_guild_logs",
        "parameters": [
          {
            "name": "guild_id",
            "in": "path",
            "description": "Guild ID or User Bot runtime scope ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Event stream containing runtime log entries",
            "content": {
              "text/event-stream": {}
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GuildResponse": {
        "type": "object",
        "description": "Guild info exposed by the API.",
        "required": ["id", "name", "permissions"],
        "properties": {
          "icon": {
            "type": ["string", "null"]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "LogEntry": {
        "type": "object",
        "description": "A log entry captured from the runtime.",
        "required": ["timestamp", "level", "target", "message"],
        "properties": {
          "guild_id": {
            "type": ["string", "null"],
            "description": "Guild ID if applicable."
          },
          "level": {
            "type": "string",
            "description": "Log level (trace, debug, info, warn, error)."
          },
          "message": {
            "type": "string",
            "description": "Log message."
          },
          "target": {
            "type": "string",
            "description": "Target/module that produced the log."
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "Timestamp in milliseconds since Unix epoch."
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "detail", "instance", "request_id"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Actionable explanation of this specific failure occurrence."
          },
          "instance": {
            "type": "string",
            "description": "Request path identifying where this failure occurred."
          },
          "request_id": {
            "type": "string",
            "description": "Correlation identifier also returned in the `X-Request-ID` header."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code returned for this failure occurrence.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Stable, human-readable summary of the failure category."
          },
          "type": {
            "type": "string",
            "description": "Stable relative URI identifying the category of failure."
          },
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemViolation"
            },
            "description": "Field-level validation failures, when applicable."
          }
        }
      },
      "ProblemViolation": {
        "type": "object",
        "required": ["field", "message"],
        "properties": {
          "field": {
            "type": "string",
            "description": "Identifies the invalid request field using its JSON or parameter path."
          },
          "message": {
            "type": "string",
            "description": "Explains how to correct the invalid field."
          }
        }
      },
...
    }
  },
...
}